tests/requirements.txt
# Testing dependencies for narko
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-benchmark>=4.0.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.0.0  # For parallel test execution

# For performance testing
psutil>=5.9.0
memory-profiler>=0.60.0

# For API testing
responses>=0.23.0
requests-mock>=1.10.0

# For security testing
bandit>=1.7.0

# Test reporting
pytest-html>=3.1.0
pytest-json-report>=1.5.0

# Coverage reporting
coverage>=7.0.0

# For mocking and testing utilities
factory-boy>=3.2.0
faker>=18.0.0